:poop: Added config file reading (full of bugs, please report)

+++ include/utils.h
@@ -3,10 +3,11 @@
 
     #include <stdbool.h>
     #include <stdio.h>
+    #include <sys/types.h>
 
 int error_severity(const char *line);
 bool is_object_file(const char *filepath);
-int get_terminal_size(void);
+ulong get_terminal_size(void);
 char *get_config_file(void);
 bool is_file_stream_null(const FILE *file_stream, const char *error_message);
 long get_file_size(const char *filepath);
